Saas-llm/backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Data validation and settings
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP requests
requests==2.31.0
httpx==0.25.2

# Database and Vector DB
chromadb==1.1.0
sqlite-utils==3.35.2

# PDF processing
PyMuPDF==1.23.8
pdfplumber==0.10.3

# ML and Embeddings
sentence-transformers==2.2.2
numpy==1.24.3

# Environment and configuration
python-dotenv==1.0.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-multipart

# CORS and middleware
python-multipart

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
